Simplify photo gallery with traditional client-side modal
Replace Next.js parallel/intercepting routes with simple client-side modal.
This fixes slow loading, broken close button, and sluggish navigation.

Changes:
- Remove complex @modal and photo/[photoId] route structure
- Create single PhotoGallery component with integrated modal
- Use client-side state instead of routing for modal
- Add redirect from old /photography/photo/:id routes
- Remove unnecessary layout.tsx and PhotoGrid.tsx

Benefits:
- Instant modal open/close
- Fast photo navigation
- Simpler codebase (-217 lines)
- Better user experience
