--- app/photography/layout.tsx
@@ -1,16 +1,0 @@
-import { ReactNode } from 'react';
-
-export default function PhotographyLayout({
-  children,
-  modal
-}: {
-  children: ReactNode;
-  modal: ReactNode;
-}) {
-  return (
-    <>
-      {children}
-      {modal}
-    </>
-  );
-}
