--- app/photography/photo/[photoId]/PhotoPage.tsx
@@ -1,126 +1,0 @@
-'use client';
-
-import { useEffect, useState } from 'react';
-import Link from 'next/link';
-import { useRouter } from 'next/navigation';
-
-import { PhotoType } from '../../getPhotos';
-
-export function PhotoPage({
-  photo,
-  photos,
-  currentIndex
-}: {
-  photo: PhotoType;
-  photos: PhotoType[];
-  currentIndex: number;
-}) {
-  const router = useRouter();
-  const [isFullScreen, setIsFullScreen] = useState(true);
-  const [imageLoaded, setImageLoaded] = useState(false);
-
-  // Reset imageLoaded when photo changes
-  useEffect(() => {
-    setImageLoaded(false);
-  }, [photo.id]);
-
-  useEffect(() => {
-    const handleKeyDown = (e: KeyboardEvent) => {
-      if (e.key === 'Escape') {
-        router.push('/photography');
-      }
-
-      if (e.key === 'ArrowLeft' && currentIndex > 0) {
-        router.push(`/photography/photo/${currentIndex - 1}`);
-      }
-      if (e.key === 'ArrowRight' && currentIndex < photos.length - 1) {
-        router.push(`/photography/photo/${currentIndex + 1}`);
-      }
-    };
-
-    window.addEventListener('keydown', handleKeyDown);
-    return () => window.removeEventListener('keydown', handleKeyDown);
-  }, [currentIndex, photos, router]);
-
-  function formatDate(dateString?: Date) {
-    if (!dateString) return '';
-
-    return new Date(dateString).toLocaleDateString('en-US', {
-      day: 'numeric',
-      month: 'long',
-      year: 'numeric'
-    });
-  }
-
-  return (
-    <div className="fixed inset-0 z-50 h-screen w-screen bg-slate-950 p-4 md:p-8">
-      <Link
-        href="/photography"
-        className="fixed top-5 right-5 z-20 grid h-10 w-10 place-items-center rounded-full bg-black/50 text-white backdrop-blur-sm hover:bg-white/90 hover:text-black"
-      >
-        ✕
-      </Link>
-      <div className="mx-auto flex h-full w-full flex-col items-center justify-center">
-        {!imageLoaded && (
-          <div className="absolute inset-0 z-10 flex items-center justify-center">
-            <div className="flex flex-col items-center gap-4">
-              <div className="h-12 w-12 animate-spin rounded-full border-4 border-slate-700 border-t-white" />
-              <p className="text-sm text-slate-400">Loading photo...</p>
-            </div>
-          </div>
-        )}
-        <img
-          src={photo.src}
-          srcSet={photo.srcSet}
-          sizes="(max-width: 768px) 100vw, (max-width: 1200px) 95vw, 2400px"
-          alt={photo.alt}
-          width={photo.width}
-          height={photo.height}
-          loading="eager"
-          decoding="async"
-          className={`cursor-pointer object-contain transition-opacity duration-300 ${
-            isFullScreen ? 'fixed h-screen w-screen object-cover' : 'max-h-[85vh] w-auto rounded-lg'
-          } ${imageLoaded ? 'opacity-100' : 'opacity-0'}`}
-          role="button"
-          onClick={() => setIsFullScreen(!isFullScreen)}
-          onLoad={() => setImageLoaded(true)}
-          style={{
-            backgroundImage: `url(${photo.blurDataURL})`,
-            backgroundSize: 'cover'
-          }}
-        />
-      </div>
-      <div className="fixed bottom-5 left-0 flex w-full items-center justify-between gap-4 px-5">
-        {currentIndex > 0 ? (
-          <Link
-            href={`/photography/photo/${currentIndex - 1}`}
-            className="m-auto grid h-10 w-10 place-items-center rounded-full bg-black/50 text-sm text-gray-200 backdrop-blur-sm hover:bg-white/90 hover:text-black"
-            aria-label="Previous photo"
-          >
-            ←
-          </Link>
-        ) : (
-          <div className="m-auto h-10 w-10" />
-        )}
-        <div className="rounded-md bg-black/50 p-2 text-center text-sm text-gray-200 backdrop-blur-sm">
-          <div className="md:hidden">{photo.location}</div>
-          <div className="md:hidden">{formatDate(photo.createdAt)}</div>
-          <div className="hidden md:block">
-            {photo.location} • {formatDate(photo.createdAt)}
-          </div>
-        </div>
-        {currentIndex < photos.length - 1 ? (
-          <Link
-            href={`/photography/photo/${currentIndex + 1}`}
-            className="m-auto grid h-10 w-10 place-items-center rounded-full bg-black/50 text-sm text-gray-200 backdrop-blur-sm hover:bg-white/90 hover:text-black"
-            aria-label="Next photo"
-          >
-            →
-          </Link>
-        ) : (
-          <div className="m-auto h-10 w-10" />
-        )}
-      </div>
-    </div>
-  );
-}
